Fix treasury section header background & refactor its hook
#214
