+++ src/screen/TreasuryScreen.tsx
@@ -1,22 +1,17 @@
-import React, {useContext} from 'react';
-import {Card, Layout, Spinner, Text, useTheme} from '@ui-kitten/components';
-import globalStyles, {standardPadding} from 'src/styles';
-import {ChainApiContext} from 'context/ChainApiContext';
-import {SectionList, StyleSheet, View} from 'react-native';
-import {u8aToString} from '@polkadot/util';
-import {ApiPromise} from '@polkadot/api';
-import {AccountId} from '@polkadot/types/interfaces';
 import Identicon from '@polkadot/reactnative-identicon';
-import {getAccountsIdentityInfo} from 'src/api/queryFunctions/getAccountsIdentityInfo';
-import {EmptyView} from 'presentational/EmptyView';
-import {useFormatBalance} from 'src/api/hooks/useFormatBalance';
-import {useQuery} from 'react-query';
-import SafeView, {noTopEdges} from 'presentational/SafeView';
+import {u8aToString} from '@polkadot/util';
+import {Card, Layout, Spinner, Text} from '@ui-kitten/components';
 import {Account} from 'layout/Account';
+import {EmptyView} from 'presentational/EmptyView';
 import Padder from 'presentational/Padder';
+import SafeView, {noTopEdges} from 'presentational/SafeView';
+import React from 'react';
+import {SectionList, StyleSheet, View} from 'react-native';
+import {useFormatBalance} from 'src/api/hooks/useFormatBalance';
+import {useTreasuryInfo} from 'src/api/hooks/useTreasuryInfo';
+import globalStyles, {standardPadding} from 'src/styles';
 
 export function TreasuryScreen() {
-  const theme = useTheme();
   const {isLoading, data, refetch} = useTreasuryInfo();
 
   const groupedData = [
@@ -47,6 +42,7 @@ export function TreasuryScreen() {
               const proposer = accountInfo?.info
                 ? u8aToString(accountInfo.info.display.asRaw)
                 : accountInfo?.accountId.toString() ?? 'unknown';
+
               return (
                 <Card style={styles.card}>
                   <View style={styles.row}>
@@ -59,7 +55,7 @@ export function TreasuryScreen() {
                     </View>
                   </View>
                   <View style={styles.row}>
-                    <Text>beneficiary: </Text>
+                    <Text category="c1">beneficiary: </Text>
                     <Account id={item.proposal.beneficiary.toString()}>
                       {({info, accountId}) => (
                         <View style={[styles.row, styles.accountsRow]}>
@@ -83,7 +79,7 @@ export function TreasuryScreen() {
             }}
             renderSectionHeader={({section: {title, data}}) => {
               return (
-                <View style={[styles.header, {backgroundColor: theme['color-basic-100']}]}>
+                <View style={styles.header}>
                   <Text category={'s1'}>{title}</Text>
                   <Text category={'p2'}>{`${data.length}`}</Text>
                 </View>
@@ -100,7 +96,7 @@ export function TreasuryScreen() {
 const styles = StyleSheet.create({
   accountsRow: {flex: 1, marginRight: 20},
   sectionList: {
-    padding: standardPadding,
+    padding: standardPadding * 2,
   },
   card: {
     marginBottom: standardPadding,
@@ -108,7 +104,7 @@ const styles = StyleSheet.create({
   beneficiary: {
     flex: 1,
   },
-  header: {justifyContent: 'space-between', flexDirection: 'row', padding: standardPadding * 2},
+  header: {justifyContent: 'space-between', flexDirection: 'row', paddingVertical: standardPadding},
   name: {marginLeft: 10, flex: 1},
   row: {
     flexDirection: 'row',
@@ -116,28 +112,3 @@ const styles = StyleSheet.create({
   },
   itemRight: {flexDirection: 'row', justifyContent: 'center', alignItems: 'center', paddingLeft: standardPadding},
 });
-
-function useTreasuryInfo() {
-  const {api} = useContext(ChainApiContext);
-
-  return useQuery(['treasury_info'], () => (api ? getTreasuryInfo(api) : undefined));
-}
-
-async function getTreasuryInfo(api: ApiPromise) {
-  const proposals = await api.derive.treasury.proposals();
-  const accountIds: AccountId[] = [];
-
-  for (const p of proposals.proposals) {
-    if (!accountIds.includes(p.proposal.proposer)) {
-      accountIds.push(p.proposal.proposer);
-    }
-  }
-  for (const p of proposals.approvals) {
-    if (!accountIds.includes(p.proposal.proposer)) {
-      accountIds.push(p.proposal.proposer);
-    }
-  }
-  const accountInfos = await getAccountsIdentityInfo(api, accountIds);
-
-  return {proposals: proposals, accountInfos: accountInfos};
-}
